Move SourcesContainer files to lab2
