+++ apps/src/lab2/utils/getInitialSources.ts
